+++ src/main/java/com/workitout/repository/MediaRepository.java
@@ -7,4 +7,4 @@ import org.springframework.data.repository.CrudRepository;
  *
  * @author Aintech
  */
-public interface MediaRepository extends CrudRepository<Media, Integer> {}
+public interface MediaRepository extends CrudRepository<Media, Long> {}
