+++ src/main/java/com/workitout/repository/RoundRepository.java
@@ -7,4 +7,4 @@ import org.springframework.data.repository.CrudRepository;
  *
  * @author Yaremchuk E.N. (aka Aintech)
  */
-public interface RoundRepository extends CrudRepository<Round, Integer> {}
+public interface RoundRepository extends CrudRepository<Round, Long> {}
