+++ src/main/java/com/workitout/repository/WorkoutRepository.java
@@ -7,4 +7,4 @@ import org.springframework.data.repository.CrudRepository;
  *
  * @author Yaremchuk E.N. (aka Aintech)
  */
-public interface WorkoutRepository extends CrudRepository<Workout, Integer> {}
+public interface WorkoutRepository extends CrudRepository<Workout, Long> {}
